requirements.txt
PyQt6>=6.5.0
Pillow>=9.0.0
numpy>=1.21.0